# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/copy": {
            "post": {
                "description": "将七牛云存储空间中的文件从一个位置复制到另一个位置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "复制文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "源文件名",
                        "name": "srcObject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "目标文件名",
                        "name": "destObject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否强制覆盖目标文件（true/false，默认为 false）",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文件复制成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "srcKey、destKey 缺失或 force 参数无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "文件复制失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/delete": {
            "delete": {
                "description": "根据文件名删除七牛云中的文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "objectName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文件删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "缺少必要参数 objectName",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "文件删除失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/download": {
            "get": {
                "description": "根据文件名生成私有或公共的下载链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "生成文件下载链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "objectName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问类型 ('public' 或 'private')",
                        "name": "accessType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "生成下载链接成功，返回下载链接",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "缺少必要参数 objectName",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/list": {
            "get": {
                "description": "列出七牛云存储空间中的文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "获取文件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件名前缀筛选条件",
                        "name": "prefix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "游标，继续从上次读取的标记处开始列出",
                        "name": "marker",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每次列举的最大文件数量 (1-1000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文件列表获取成功，返回文件信息及下一页游标",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "文件列表获取失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/move": {
            "post": {
                "description": "将七牛云存储空间中的文件从一个位置移动到另一个位置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "移动文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "源文件名",
                        "name": "srcObject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "目标文件名",
                        "name": "destObject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否强制覆盖目标文件（true/false，默认为 false）",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文件移动成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "srcKey、destKey 缺失或 force 参数无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "文件移动失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/upload": {
            "get": {
                "description": "根据文件路径和目标对象名称，将文件上传至七牛云存储",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "上传文件至七牛云",
                "parameters": [
                    {
                        "type": "string",
                        "description": "本地文件路径",
                        "name": "filePath",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "目标对象名称",
                        "name": "objectName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功，返回文件信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "缺少必要参数 filePath 和 objectName",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "上传失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}